shell/telnet: fix uint8_t truncation that wedged multi-segment output
tiku_shell_io_tcp_flush() declared the per-call send length as uint8_t but assigned it tx_pos (uint16_t).  Once a command's buffered output passed 255 bytes the length wrapped: for help (~1898 B) the low byte walked 106 -> 18 -> 0, and tcp_send(len=0) returns an error every poll, so the drain wedged after 2 segments (88+18=106 bytes) and the session appeared frozen.  Output <=255 B (e.g. the 244-byte banner) slipped under it, which is why connect/echo worked but help/info/ps froze.

Widen chunk to uint16_t.  HW-verified apollo4l: help now drains the full 1898 bytes with prompt, repeatedly; telnet multi-cmd 11/11, disconnect 9/9, RST 7/7, idle 5/5 (all the freeze regressions clear).

Found by J-Link-reading tx_buf counters: the full output reached tx_buf intact (0 drops, high-water 1898) but only 106 bytes were ever sent.

+++ kernel/shell/tiku_shell_io_tcp.c
@@ -137,7 +137,7 @@ static uint8_t last_was_cr;
 void
 tiku_shell_io_tcp_flush(void)
 {
-    uint8_t chunk;
+    uint16_t chunk;
     uint16_t mss;
 
     if (!telnet_conn || tx_pos == 0) {
@@ -154,8 +154,13 @@ tiku_shell_io_tcp_flush(void)
     mss = telnet_conn->snd_mss;
     chunk = tx_pos;
     if (chunk > mss) {
-        chunk = (uint8_t)mss;
+        chunk = mss;
     }
+    /* chunk and tx_pos are both uint16_t: a command's output can exceed 255
+     * bytes (e.g. `help` is ~1.9 KB), so a uint8_t here would wrap -- once
+     * tx_pos passed 0xFF the low byte could land on 0 and tcp_send(len=0)
+     * returns an error every poll, wedging the drain a couple of segments in
+     * (the banner, <256 B, slipped under it).  Keep the full 16-bit length. */
 
     if (tiku_kits_net_tcp_send(telnet_conn,
                                tx_buf, chunk) != TIKU_KITS_NET_OK) {
